Init MessageCachePerformance extension
Due to https://phabricator.wikimedia.org/T193271, lookups for messages that are
not defined in core/extensions and are not customized on the wiki trigger
memcached/APCu lookups. This can be quite expensive when many wikis and messages
are involved.

This implements an extension that detects the most common messages fitting this
pattern, and replaces them with a blank message defined in code to eliminate
unwanted I/O.