Set messages as nonexistent instead of using a blank message key
The current logic that reroutes messages known to not exist to use a blank
system message under the hood is incompatible with some existing code paths
(e.g. LanguageConverter) that rely on a Message::exists() existence check to
trigger fallback logic. Instead, let's return false from the hook handler for
message keys that are known to not exist to avoid expensive MessageCache
operations while maintaining compatibility with code that relies on message
existence checks to function correctly. This is made possible by
https://github.com/Wikia/mediawiki/pull/19.