PLATFORM-7961 | Upgrade MessageCachePerformance to work with MW 1.39
Extension was working fine under new version but there was bug (or rather unexpected feature), that message string where not only matched by the prefix but also in the middle. Therefore, removed aho-corasick and replaced with regex_match. Also, added the posibility to return debug info with skipped message into the output, which can be enabled by the variable.